+++ gems/kettle-jem/CHANGELOG.md
@@ -51,6 +51,7 @@ Please file a bug if you notice a violation of semantic versioning.
 ### Changed
 
 - Gem templates now require `appraisal2` >= 3.1.3.
+- Generated style Gemfiles now require `appraisal2-rubocop` >= 0.2.2.
 - Generated main Gemfiles now require `nomono` >= 1.0.6.
 - Generated gemspecs now require `version_gem` >= 1.1.13.
 - Generated gemspecs now require `gitmoji-regex` >= 2.0.3.